+++ nimwcpkg/emails/_senderrors_admins.nim
@@ -2,6 +2,5 @@
 
 proc sendEmailAdminError*(msg: string) {.async.} =
   ## Send email - user removed
-  preconditions msg.len > 0
   await sendAdminMailNow("Admin: Error occurred", genEmailMessage(
     adminErrorMsg.format(msg, msg.countLines, now())))