+++ nimwcpkg/plugins/_settings.nim
@@ -4,8 +4,6 @@ proc extensionSettings(): seq[string] =
   ## Proc to check if plugins listed in plugins_imported.txt
   ## are enabled or disabled. The result will be "true:pluginname"
   ## or "false:pluginname".
-  preconditions dirExists"plugins/", fileExists"plugins/plugin_import.txt"
-  postconditions if readFile("plugins/plugin_import.txt").splitLines.len > 1: result.len > 0 else: true
   let plugins = readFile("plugins/plugin_import.txt").splitLines
 
   # Walk through files and folders in the plugin directory