+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v6.1.3
+## Changes
+- Deprecate libravatar package
+- Deprecate contra package
+- Fix nim v1.6.x
+
 ## v6.0.3
 ## Changes
 - Fix #121, #122 and #123